// angular-src/src/app/swagger-api/model/registerParams.ts
/**
 * JAANartc
 * JAANartc OpenAPI Documentations
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface RegisterParams {
    userName: string;
    fullName?: string;
    password: string;
}